+++ src/viewpanel/ViewPanel.tsx
@@ -1,20 +1,15 @@
 import * as React from 'react';
-import { CommandToCode } from './CommandToCode';
 import { Actions } from './components/Actions';
 import { BaseView } from './components/BaseView';
 import { Collapsible } from './components/Collapsible';
 import { GlobalSettings } from './components/GlobalSettings';
-import { BugIcon } from './components/Icons/BugIcon';
-import { FileIcon } from './components/Icons/FileIcon';
-import { FolderOpenedIcon } from './components/Icons/FolderOpenedIcon';
 import { ListUnorderedIcon } from './components/Icons/ListUnorderedIcon';
-import { SettingsIcon } from './components/Icons/SettingsIcon';
 import { SymbolKeywordIcon } from './components/Icons/SymbolKeywordIcon';
 import { TagIcon } from './components/Icons/TagIcon';
+import { OtherActions } from './components/OtherActions';
 import { SeoStatus } from './components/SeoStatus';
 import { Spinner } from './components/Spinner';
 import { TagPicker } from './components/TagPicker';
-import { MessageHelper } from './helper/MessageHelper';
 import useMessages from './hooks/useMessages';
 import { TagType } from './TagType';
 
@@ -36,18 +31,6 @@ export const ViewPanel: React.FunctionComponent<IViewPanelProps> = (props: React
     );
   }
 
-  const openSettings = () => {
-    MessageHelper.sendMessage(CommandToCode.openSettings);
-  };
-  
-  const openFile = () => {
-    MessageHelper.sendMessage(CommandToCode.openFile);
-  };
-  
-  const openProject = () => {
-    MessageHelper.sendMessage(CommandToCode.openProject);
-  };
-
   return (
     <div className="frontmatter">
       <div className={`ext_actions`}>
@@ -60,7 +43,7 @@ export const ViewPanel: React.FunctionComponent<IViewPanelProps> = (props: React
           settings && metadata && <Actions metadata={metadata} settings={settings} />
         }
 
-        <Collapsible title="Metadata" className={`absolute w-full`}>
+        <Collapsible title="Metadata" className={`inherit z-20`}>
           {
             <TagPicker type={TagType.keywords} 
                        icon={<SymbolKeywordIcon />}
@@ -94,24 +77,8 @@ export const ViewPanel: React.FunctionComponent<IViewPanelProps> = (props: React
             )
           }
         </Collapsible>
-      </div>
 
-      <div className={`ext_settings`}>
-        <div className="ext_link_block">
-          <button onClick={openSettings}><SettingsIcon /> Open settings</button>
-        </div>
-
-        <div className="ext_link_block">
-          <button onClick={openFile}><FileIcon /> Reveal file in folder</button>
-        </div>
-
-        <div className="ext_link_block">
-          <button onClick={openProject}><FolderOpenedIcon /> Reveal project folder</button>
-        </div>
-
-        <div className="ext_link_block">
-          <a href="https://github.com/estruyf/vscode-front-matter/issues" title="Open an issue on GitHub"><BugIcon /> Report an issue</a>
-        </div>
+        <OtherActions isFile={true} />
       </div>
     </div>
   );